LRCI-2295 Move StopWatchRecord & StopWatchRecordsGroup to their own classes
